+++ tcvectordb/helper.go
@@ -1,6 +1,9 @@
 package tcvectordb
 
 import (
+	"encoding/json"
+	"log"
+
 	"github.com/tencent/vectordatabase-sdk-go/tcvectordb/olama"
 )
 
@@ -17,6 +20,8 @@ func ConvertDbType(dataType olama.DataType) string {
 
 func ConvertField2Grpc(field *Field) (result *olama.Field) {
 	switch field.Type() {
+	case Double:
+		result = &olama.Field{OneofVal: &olama.Field_ValDouble{ValDouble: field.Float()}}
 	case Uint64:
 		result = &olama.Field{OneofVal: &olama.Field_ValU64{ValU64: field.Uint64()}}
 	case String:
@@ -28,6 +33,13 @@ func ConvertField2Grpc(field *Field) (result *olama.Field) {
 			byteArray = append(byteArray, []byte(s))
 		}
 		result = &olama.Field{OneofVal: &olama.Field_ValStrArr{ValStrArr: &olama.Field_StringArray{StrArr: byteArray}}}
+	case Json:
+		jsonData, err := json.Marshal(field.Val)
+		if err != nil {
+			log.Printf("[Error] marshal failed when converting field to rpc request body. err: %v", err.Error())
+			return
+		}
+		result = &olama.Field{OneofVal: &olama.Field_ValJson{ValJson: jsonData}}
 	}
 	return
 }
@@ -43,6 +55,13 @@ func ConvertGrpc2Field(field *olama.Field) (result *Field) {
 		result.Val = v.ValDouble
 	case *olama.Field_ValStrArr:
 		result.Val = ConvertByte2StringSlice(v.ValStrArr.StrArr)
+	case *olama.Field_ValJson:
+		result.Val = make(map[string]interface{}, 0)
+		err := json.Unmarshal(v.ValJson, &result.Val)
+		if err != nil {
+			log.Printf("[Error] Unmarshal failed when converting rpc request body to field. err: %v", err.Error())
+			return
+		}
 	}
 	return
 }