+++ tcvectordb/rpc_base_document.go
@@ -700,7 +700,9 @@ func (r *rpcImplementerFlatDocument) HybridSearch(ctx context.Context, databaseN
 	req.Search.Filter = params.Filter.Cond()
 	req.Search.RetrieveVector = params.RetrieveVector
 	req.Search.Outputfields = params.OutputFields
-	req.Search.Limit = uint32(*params.Limit)
+	if params.Limit != nil {
+		req.Search.Limit = uint32(*params.Limit)
+	}
 
 	res, err := r.rpcClient.HybridSearch(ctx, req)
 	if err != nil {
@@ -961,4 +963,12 @@ func (r *rpcImplementerFlatDocument) Count(ctx context.Context, databaseName, co
 		return nil, err
 	}
 	return &CountDocumentResult{Count: res.Count}, nil
+}
+
+func (r *rpcImplementerFlatDocument) UploadFile(ctx context.Context, databaseName, collectionName string, param UploadFileParams) (result *UploadFileResult, err error) {
+	return uploadFile(ctx, r.SdkClient, databaseName, collectionName, param)
+}
+
+func (r *rpcImplementerFlatDocument) GetImageUrl(ctx context.Context, databaseName, collectionName string, param GetImageUrlParams) (result *GetImageUrlResult, err error) {
+	return getImageUrl(ctx, r.SdkClient, databaseName, collectionName, param)
 }